fix: global logrus configuration (#575)
Use ConfigureLogging to globally configure logrus so that all logs are
bound to the same configuration, e.g. level and formatter. When create
a new logger, such as in Mailer, use logrus.StandardLogger() instead of
logrus.New() so that it inherits the global configuration

Also, provide pop with a logging function that wraps logrus. To be
consistent with the current behavior, a new setting, GOTRUE_LOG_SQL was
added that takes the values of "all", "statement" or "none" (default to
"none"). "statement" will log any SQL statement that goes through POP
while "all" will log the statement + arguments. "all" could log
sensitive information.

Co-authored-by: Kang Ming <kang.ming1996@gmail.com>