feat(saml): add SSO authorization API (disabled by default) (#786)
Adds the SSO authorization API with which you can initiate a single
sign-on flow.

You can do this by sending a POST request to `/sso`. The parameters are
as follows:

```json
{
  "provider_id": "cb059692-59ec-4ac3-848a-0d6b48569eac"
}
```

Which will initiate SSO with a known provider ID.

```json
{
  "domain": "example.com"
}
```

Which will initiate SSO by looking up the registered domain for a
provider. This is the more useful version of the call.

The call results in a [303 See
Other](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/303)
response that redirects to the provider's authorization URL.

[PR for gotrue-js.](https://github.com/supabase/gotrue-js/pull/526)