feat: internalize implementation (#925)
Most of the GoTrue Go code should not be used in external imports, as
it's an internal implementation. It is thus moved to an
[`internal`](https://go.dev/doc/go1.4#internalpackages) package which
can't be imported from the outside.

From this point on you won't be able to access implementation details by
importing `github.com/netlify/gotrue/api` for example.