feat: make dropping `users_email_key` backward compatible (#995)
Dropping the `users_email_key` can sometimes be an issue as the
`auth.users.email` column has been used as a foreign key reference in
certain projects. This PR will continue with the migration, but will
raise a notice that the developers of such projects should take care of
this change manually. If it is not handled, SSO will not work and future
compatibility with new versions may be an issue.