ci: add conventional commits linter (#1392)
Adds a linter for commits and PRs to ensure that correct structure is
observed. Certain directives are not compatible with this project:

- Breaking changes are not allowed. Therefore PR titles or commit
messages must not use the `!` nor the `BREAKING CHANGE` directive in
descriptions or commit messages.
- The `Release-As` directive is not supported as it can break the
release process.
- For cleanliness, only a few conventional commit prefixes are allowed.

These are linted using a simple JavaScript script that runs on PRs and
pushes.